dvc: optimize imports
Fixes #903
